+++ ase/calculators/calculator.py
@@ -107,10 +107,10 @@ all_changes = ['positions', 'numbers', 'cell', 'pbc',
 names = ['abinit', 'ace', 'aims', 'amber', 'asap', 'castep', 'cp2k',
          'crystal', 'demon', 'dftb', 'dftd3', 'dmol', 'eam', 'elk',
          'emt', 'espresso', 'exciting', 'ff', 'fleur', 'gaussian',
-         'gpaw', 'gromacs', 'gulp', 'hotbit', 'jacapo', 'lammpslib',
-         'lammpsrun', 'lj', 'mopac', 'morse', 'nwchem', 'octopus',
-         'onetep', 'openmx', 'psi4', 'qchem', 'siesta', 'tip3p',
-         'tip4p', 'turbomole', 'vasp']
+         'gpaw', 'gromacs', 'gulp', 'hotbit', 'jacapo', 'kim',
+         'lammpslib', 'lammpsrun', 'lj', 'mopac', 'morse', 'nwchem',
+         'octopus', 'onetep', 'openmx', 'psi4', 'qchem', 'siesta',
+         'tip3p', 'tip4p', 'turbomole', 'vasp']
 
 
 special = {'cp2k': 'CP2K',